+++ lib/features/home/domain/repository/category_repository.dart
@@ -3,5 +3,5 @@ import 'package:flutter_vtv/core/constants/typedef.dart';
 import '../entities/category_entity.dart';
 
 abstract class CategoryRepository {
-  RespEitherData<List<CategoryEntity>> getAllParentCategories();
+  FRespData<List<CategoryEntity>> getAllParentCategories();
 }