+++ lib/features/home/domain/repository/product_repository.dart
@@ -2,5 +2,5 @@ import '../../../../core/constants/typedef.dart';
 import '../dto/product_dto.dart';
 
 abstract class ProductRepository {
-  RespEitherData<ProductDTO> getSuggestionProductsRandomly(int page, int size);
+  FRespData<ProductDTO> getSuggestionProductsRandomly(int page, int size);
 }